Extract model intro markers to custom hook
Move marker creation and tracking logic from the context provider to a dedicated useModelIntroMarkers hook. This improves separation of concerns and makes the hook's behavior easier to test independently. Update Conversation and ConversationEntries components to use the hook and accept markers as props instead of accessing them from context.
